backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
httpx==0.25.2
python-dotenv==1.0.0
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
